@vkostyukov 2017-01-14
If we still want to allow people seeing pretty JSON, I'd probably vote for adding `jacksonPretty*` variants along with `jacksonPrint*` methods that explicitly use pretty printer.