@travisbrown 2017-01-14
👍, I think there's a good argument for doing this apart from the possible performance benefits.