@vkostyukov 2017-01-16
@travisbrown I think so too! Hope that will make it to the Circe 0.7.0 release.