@travisbrown 2017-01-16
@vkostyukov Yep—I've got one iteratee.io thing to fix first but am hoping to have the circe release out this afternoon.